Update release.rb
I am trying to cover a case here: Say the GIT tagging process went through without any issues, but the SDK publish step failed for some reason. So the next time you restart the job, the tagging step will fail as there is TAG with the version already. I am not sure if deleting a tag is a good way to do things. Please review.